added sh, pixelscript inits

+++ inits/filter.frag
@@ -1,6 +1,6 @@
 
 #ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif
 
 uniform float u_time;
@@ -11,6 +11,6 @@ uniform sampler2D u_tex0;
 void main (void) {
     vec2 st = gl_FragCoord.xy/u_resolution;
     vec4 source = texture2D( u_tex0, st );
+    
     gl_FragColor = vec4( st.x*0.2, source.g, source.b, 1.0 ); 
-    // source.a
 }
